+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to paopao-ce are documented in this file.
 ## 0.5.0+dev ([`dev`](https://github.com/rocboss/paopao-ce/tree/dev))
+### Changed
+- web frontend remove pagination to display content and use load more mechanism load page content.
+
 ## 0.4.0
 ### Added
 - add `pprof` feature support [#327](https://github.com/rocboss/paopao-ce/pull/327)